% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_reference_lists}
\alias{compare_reference_lists}
\title{Compare reference lists}
\usage{
compare_reference_lists(provided_refs, extracted_refs)
}
\arguments{
\item{provided_refs}{data frame with columns \code{doi}, \code{title}, \code{author}, \code{year}
and \code{journal}. This dataframe will typically be the user input e.g. a list of
references that satisfy the inclusion criteria in a systematic review.}

\item{extracted_refs}{a data frame with columns \code{doi}, \code{title}, \code{author}, \code{year}
and \code{journal}. This dataframe will typically be the reference list extracted
from a set of papers e.g. using \code{\link{extract_references}}.}
}
\description{
Compare two lists of references to identify references that are in one list
but not the other
}
\details{
Compares two lists of references to identify references that are in
one list but not the other. We first check if any dois in \code{extracted_refs} are
in \code{provided_refs}. Then we check if any titles in \code{extracted_refs} are in
\code{provided_refs}. Titles are compared after pre-processing them with
\code{\link{cleanup_strings}}. If the list \code{extracted_refs} has been generated
by \code{\link{extract_references}}, the titles will already be pre-processed.
}
\seealso{
\code{\link{extract_references}}, \code{\link{cleanup_strings}}
}
\author{
Sangeeta Bhatia
}
