% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deduplicate_references}
\alias{deduplicate_references}
\title{Deduplicate a list of references}
\usage{
deduplicate_references(x)
}
\arguments{
\item{x}{data.frame of references. Must have columns "doi" and "title".}
}
\value{
a deduplicated data.frame of references with an additional column
called "cleanup_title" that contains the cleaned up title string. See
\code{\link{cleanup_strings}} for details.
}
\description{
Identify duplicates based on DOI and title. Remove duplicates and return a
deduplicated list of references.
}
\author{
Sangeeta Bhatia
}
