% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_missing_article_info_}
\alias{get_missing_article_info_}
\title{Retrieve additional information for multiple references using crossref}
\usage{
get_missing_article_info_(ref_list, method)
}
\arguments{
\item{ref_list}{A list of references, most likely output from
\code{\link{extract_references}}. If you have produced a list of references
manually, ensure that it has the following columns: doi, title, authors.}

\item{method}{retrieval backend; one of "crossref" or "openalex"}
}
\value{
the input data.frame with DOI updated where it could be retrirved
from rcrossref
}
\description{
This function takes a data.frame of references and queries crossref for
additional information for each reference that is missing a DOI. It returns
a data.frame with the
title and DOI information for the input query parameters.
}
\author{
Sangeeta Bhatia
}
