% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_missing_article_info}
\alias{get_missing_article_info}
\title{Retrieve additional information for references using crossref}
\usage{
get_missing_article_info(title, authors)
}
\arguments{
\item{title}{}

\item{x}{data.frame of references. Must have columns "doi" and "title".}
}
\description{
Automated extraction leaves a lot wanting. The only thing we can reliably
extract is the title. We can use this title to query crossref and get
additional information. This function takes a data.frame of references and
queries crossref for additional information. It returns a data.frame with
}
\author{
Sangeeta Bhatia
}
